Refactors the renewing OAuth Provider into a separate class
