+++ src/main/java/tech/redroma/yelp/OAuthTokenProvider.java
@@ -135,7 +135,7 @@ public interface OAuthTokenProvider
             .usingMessage("Alchemy HTTP cannot be null")
             .is(notNull());
 
-        return new OAuthObtainer(http, authURL, clientId, clientSecret);
+        return new OAuthTokenProviderRenewing(http, authURL, clientId, clientSecret);
     }
 
     /**
@@ -156,176 +156,5 @@ public interface OAuthTokenProvider
     }
 
 
-    static final class OAuthObtainer implements OAuthTokenProvider
-    {
-
-        private final AlchemyHttp http;
-        private final URL authenticationURL;
-        private final String clientId;
-        private final String clientSecret;
-
-        private static final int BAD_REQUEST = 400;
-
-        private static final Logger LOG = LoggerFactory.getLogger(OAuthObtainer.class);
-
-        OAuthObtainer(AlchemyHttp http, URL authenticationURL, String clientId, String clientSecret)
-        {
-            checkThat(http, authenticationURL)
-                .are(notNull());
-
-            checkThat(clientId, clientSecret)
-                .are(nonEmptyString())
-                .are(stringWithLengthGreaterThan(2));
-
-            this.http = http;
-            this.authenticationURL = authenticationURL;
-            this.clientId = clientId;
-            this.clientSecret = clientSecret;
-        }
-
-        @Override
-        public String getToken()
-        {
-            JsonObject response = null;
-
-            try
-            {
-                response = http.go()
-                    .post()
-                    .nothing()
-                    .usingQueryParam(Keys.GRANT_TYPE, "client_credentials")
-                    .usingQueryParam(Keys.CLIENT_ID, clientId)
-                    .usingQueryParam(Keys.CLIENT_SECRET, clientSecret)
-                    .usingHeader(Keys.CONTENT_TYPE, Keys.URL_ENCODED)
-                    .expecting(JsonObject.class)
-                    .at(authenticationURL);
-            }
-            catch (AlchemyHttpException ex)
-            {
-                HttpResponse yelpResponse = ex.getResponse();
-                int status = yelpResponse.statusCode();
-
-                if (status == BAD_REQUEST)
-                {
-                    throw new YelpBadArgumentException("Client ID or Secret are incorrect: " + yelpResponse.body());
-                }
-
-                throw new YelpOperationFailedException("Failed to get access token.", ex);
-            }
-
-            checkResponse(response);
-            printExpiration(response);
-
-            return tryToGetTokenFrom(response);
-        }
-
-        @Override
-        public int hashCode()
-        {
-            int hash = 7;
-            hash = 79 * hash + Objects.hashCode(this.http);
-            hash = 79 * hash + Objects.hashCode(this.authenticationURL);
-            return hash;
-        }
-
-        @Override
-        public boolean equals(Object obj)
-        {
-            if (this == obj)
-            {
-                return true;
-            }
-            if (obj == null)
-            {
-                return false;
-            }
-            if (getClass() != obj.getClass())
-            {
-                return false;
-            }
-            final OAuthObtainer other = (OAuthObtainer) obj;
-            if (!Objects.equals(this.http, other.http))
-            {
-                return false;
-            }
-            if (!Objects.equals(this.authenticationURL, other.authenticationURL))
-            {
-                return false;
-            }
-            return true;
-        }
-
-        @Override
-        public String toString()
-        {
-            return "OAuthObtainer{" + "http=" + http + ", authenticationURL=" + authenticationURL + '}';
-        }
-
-        private void printExpiration(JsonObject response)
-        {
-            JsonElement expiration = response.get(Keys.EXPIRATION);
-
-            if (expiration.isJsonPrimitive())
-            {
-                JsonPrimitive expirationValue = expiration.getAsJsonPrimitive();
-
-                if (expirationValue.isNumber())
-                {
-                    int expirationSeconds = expirationValue.getAsInt();
-
-                    long expirationDays = TimeUnit.SECONDS.toDays(expirationSeconds);
-                    long expirationMinutes = TimeUnit.SECONDS.toMinutes(expirationSeconds);
-                    LOG.debug("Yelp Token expires in {} days or {} minutes", expirationDays, expirationMinutes);
-                }
-                else
-                {
-                    LOG.warn("Received unexpected token expiration: " + expirationValue);
-                }
-
-            }
-        }
-
-        private void checkResponse(JsonObject response)
-        {
-            checkThat(response)
-                .throwing(YelpOperationFailedException.class)
-                .usingMessage("Received unexpected null response")
-                .is(notNull());
-
-            checkThat(response.has(Keys.EXPIRATION))
-                .usingMessage("OAuth response is missing expiration information: " + response)
-                .is(trueStatement());
-
-            checkThat(response.has(Keys.TOKEN))
-                .usingMessage("OAUth response is missing access token: " + response)
-                .throwing(YelpOperationFailedException.class)
-                .is(trueStatement());
-        }
-
-        private String tryToGetTokenFrom(JsonObject response)
-        {
-            checkThat(response.has(Keys.TOKEN))
-                .usingMessage("Yelp AUTH response is missing the token")
-                .throwing(YelpBadArgumentException.class)
-                .is(trueStatement());
-
-            return response.get(Keys.TOKEN).getAsString();
-        }
-
-        private static class Keys
-        {
-
-            static final String GRANT_TYPE = "grant_type";
-            static final String CLIENT_ID = "client_id";
-            static final String CLIENT_SECRET = "client_secret";
-            static final String CONTENT_TYPE = "Content-Type";
-            static final String URL_ENCODED = "application/x-www-form-urlencoded";
-
-            //Response Keys
-            static final String EXPIRATION = "expires_in";
-            static final String TOKEN = "access_token";
-        }
-
-    }
 
 }
